split StreamRelationSuite
